[refactor] FrameResult 리팩토링후 불필요한 dto제거
