+++ src/test/java/bowling/domain/FramesTest.java
@@ -2,7 +2,7 @@ package bowling.domain;
 
 import bowling.domain.State.*;
 import bowling.domain.frame.*;
-import bowling.dto.FrameResult2;
+import bowling.dto.FrameResult;
 import bowling.dto.FrameResults;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -21,7 +21,7 @@ public class FramesTest extends FrameTest {
         Frames frames = Frames.init(totalNumberOfFrame);
         FrameResults result = frames.result();
 
-        List<FrameResult2> results = result.results();
+        List<FrameResult> results = result.results();
 
         assertThat(results.size()).isEqualTo(totalNumberOfFrame);
         assertThat(frames.currentFrameNumber()).isEqualTo(FrameNumber.first());
@@ -80,7 +80,7 @@ public class FramesTest extends FrameTest {
 
         FrameResults result = frames.result();
         List<String> states = result.results().stream()
-                .map(FrameResult2::state)
+                .map(FrameResult::state)
                 .collect(Collectors.toList());
 
         assertThat(states).containsExactlyInAnyOrder(STRIKE_SYMBOL, firstPinCountOfSecondFrame + SEPARATOR + secondPinCountOfSecondFrame, EMPTY_SYMBOL);
